# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "JWT access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/send_code": {
      "post": {
        "operationId": "AuthController_sendCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login and get JWT",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/company-types": {
      "post": {
        "operationId": "CompanyTypeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "b2b"
                  },
                  "name": {
                    "type": "string",
                    "example": "B2B"
                  }
                },
                "required": [
                  "code",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create company type",
        "tags": [
          "Company Types"
        ]
      },
      "get": {
        "operationId": "CompanyTypeController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of company types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "uuid"
                      },
                      "code": {
                        "type": "string",
                        "example": "b2b"
                      },
                      "name": {
                        "type": "string",
                        "example": "B2B"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2025-09-18T10:00:00.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2025-09-18T10:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Types"
        ]
      }
    },
    "/api/company-types/{id}": {
      "get": {
        "operationId": "CompanyTypeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company type by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "uuid"
                    },
                    "code": {
                      "type": "string",
                      "example": "b2b"
                    },
                    "name": {
                      "type": "string",
                      "example": "B2B"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2025-09-18T10:00:00.000Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2025-09-18T10:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Types"
        ]
      },
      "patch": {
        "operationId": "CompanyTypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "b2c"
                  },
                  "name": {
                    "type": "string",
                    "example": "B2C"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update company type",
        "tags": [
          "Company Types"
        ]
      },
      "delete": {
        "operationId": "CompanyTypeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Types"
        ]
      }
    },
    "/api/acl/permissions": {
      "get": {
        "operationId": "ACLController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all permissions",
        "tags": [
          "Access Control"
        ]
      }
    },
    "/api/acl/me/permissions": {
      "get": {
        "operationId": "ACLController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List current user permissions (role + overrides)",
        "tags": [
          "Access Control"
        ]
      }
    },
    "/api/acl/roles": {
      "get": {
        "operationId": "ACLController_listRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List roles",
        "tags": [
          "Access Control"
        ]
      },
      "post": {
        "operationId": "ACLController_createRole",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create role",
        "tags": [
          "Access Control"
        ]
      }
    },
    "/api/acl/roles/{id}": {
      "patch": {
        "operationId": "ACLController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update role",
        "tags": [
          "Access Control"
        ]
      }
    },
    "/api/acl/roles/{id}/permissions": {
      "post": {
        "operationId": "ACLController_setRolePerms",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permissionCodes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "permissionCodes"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set role permissions by codes",
        "tags": [
          "Access Control"
        ]
      }
    },
    "/api/acl/users/{id}/overrides": {
      "post": {
        "operationId": "ACLController_setUserOverrides",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "overrides": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "allowed": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "overrides"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set user permission overrides",
        "tags": [
          "Access Control"
        ]
      }
    },
    "/api/acl/users/{userId}/role": {
      "post": {
        "operationId": "ACLController_setUserRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Access Control"
        ]
      }
    },
...
  },
  "info": {
    "title": "Best CRM API",
    "description": "API documentation for Best CRM backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Иван Иванов"
          },
          "email": {
            "type": "string",
            "example": "ivan@example.com"
          }
        },
        "required": [
          "fullName",
          "email"
        ]
      },
      "AuthTokenResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "access_token"
        ]
      },
      "SendCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ivan@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ivan@example.com"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
...
      "SetUserRoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          }
        },
        "required": [
          "roleId"
        ]
      },
...
    }
  }
}